Create an Interval type
This commit introduces an Interval type with a representation derived from
mathematical literature: "[lo,hi)".  A square bracket indicates an
"inclusive" number (that is, the interval includes that number), while a
round bracket (aka parenthesis) indicates an "exclusive" number (that is,
the interval excludes that number).  The Interval type will be used to
specify the acceptable number of uses of a particular option or parameter.