Replace dependency injector
This commit replaces the dependency injector.  A flaw in the usage pattern
of the original was realized: it is necessary to determine the required
input types separately from invoking the method.  The new dependency
injector is significantly simplified by identifying a method and returning
an object with a Call method; the returned object also has a Deps field
which indicates the required dependencies.